Revert "Merge branch 'FE' into 'master'"

See merge request s09-webmobile1-sub2/S09P12A703!31

+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/ModifyMeetingRequestDTO.java
@@ -21,7 +21,7 @@ public class ModifyMeetingRequestDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private int maxParticipant;
+    private Long maxParticipant;
 
     private String title;
